Add `AsyncObjectMapper`, for asynchronous mapping functions.
Ideally, the async operations would be performed before any mapping occurs; in the real world, sometimes you need to get values "just-in-time".

Unfortunately, this is a lot of duplicate code and types. The synchronous implementation is different enough that I couldn't think of a way to reuse it.

The JSDocs are just copy/pasted, without examples, and probably somewhat lacking.